Game class created, levelview moved to view
