+++ src/App/Game.java
@@ -1,5 +1,52 @@
 package App;
 
-public class Game {
+import view.LevelTypeSelectView;
+import view.LevelView;
 
+public class Game {
+	LevelTypeSelectView ltsv;
+	//TODO add abstractLevelModel levels[]
+	//TODO add abstractLevelModel currentLevel
+	LevelView levelView;
+	//ExitLevelScreen exitLevel
+	
+	Game(){
+		
+	}
+	
+	void initialize(){
+		
+	}
+	
+	void initializeView(){
+		//TODO Initialize splash screen or level select view here?
+	}
+	
+	void initializeControllers(){
+		
+	}
+	
+	void initializeModels(){
+		
+	}
+	
+	void unlockNextLevel(int nextLevelID){
+		
+	}
+	
+	void initializeLevelView(){
+		
+	}
+	
+	void initializeLevelControllers(){
+		
+	}
+	
+	void initializeExitLevelScreen(){
+		
+	}
+	
+	void initializeExitLevelScreenControllers(){
+		
+	}
 }
